Address review suggestions; Move code around
