+++ frontend/catalyst/device/__init__.py
@@ -16,9 +16,11 @@
 Internal API for the device module.
 """
 
-from catalyst.device.qjit_device import QJITDevice, QJITDeviceNewAPI
-
-__all__ = (
-    "QJITDevice",
-    "QJITDeviceNewAPI",
+from catalyst.device.qjit_device import (
+    BackendInfo,
+    QJITDevice,
+    QJITDeviceNewAPI,
+    extract_backend_info,
 )
+
+__all__ = ("QJITDevice", "QJITDeviceNewAPI", "BackendInfo", "extract_backend_info")
