+++ frontend/test/lit/test_decomposition.py
@@ -21,7 +21,7 @@ import jax
 import pennylane as qml
 
 from catalyst import cond, for_loop, measure, qjit, while_loop
-from catalyst.utils.runtime import pennylane_operation_set
+from catalyst.utils.toml import pennylane_operation_set
 from catalyst.utils.toml import ProgramFeatures, get_device_capabilities
