+++ frontend/test/pytest/test_decomposition.py
@@ -19,7 +19,7 @@ import pytest
 from jax import numpy as jnp
 
 from catalyst import CompileError, ctrl, measure, qjit
-from catalyst.utils.runtime import pennylane_operation_set
+from catalyst.utils.toml import pennylane_operation_set
 from catalyst.utils.toml import ProgramFeatures, get_device_capabilities
